stock_market_research_data_warehouse/pyproject.toml
[project]
name = "stock_market_research_data_warehouse"
version = "0.1"
description = "Ignestion and processing pipelines for stock market research project"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "apache-airflow[celery,postgres,amazon]==3.0.3",
    "apache-airflow-providers-common-sql==1.27.4",
    "pandas==2.3.1",
    "psycopg2-binary==2.9.10",
    "pyarrow==21.0.0",
    "python-dotenv",
    "requests==2.32.4",
    "yfinance>=0.2.65,<0.3.0",
]

[dependency-groups]
dev = [
    "autoflake",
    "black",
    "flake8",
    "ipython",
    "pre-commit",
    "pycln",
    "pytest",
    "reorder-python-imports",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
env_files = ["tests/.env.test"]
log_cli = "True"
log_cli_level = "INFO"

[flake8]
ignore = 'D100,W503,E203'
max-line-length = 100
